refactor(input-validation): split runtime validator into SRP submodules (#0000)
